# docs/api/openapi.yaml
openapi: 3.0.3

info:
  title: MuzaLife Backend API
  version: 1.0.0
  description: |
    REST API for the **MuzaLife** platform — a web store for creative scenarios,
    quests, poetry, and other musical-life materials.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via `/api/auth/login` or
    one of the OAuth endpoints.  Pass it in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    ## Registration flow
    Registration is a two-step process:
    1. `POST /api/auth/register/initiate` → sends a 6-digit OTP to the email.
    2. `POST /api/auth/register/verify`   → validates OTP, creates user, returns JWT.

  contact:
    name: Vladymir Koroviakov
    email: v.korovyakov@student.sumdu.edu.ua

servers:
  - url: https://localhost:5001/api
    description: Local development (HTTPS)

tags:
  - name: Auth
    description: Registration, login, OAuth
  - name: Users
    description: User profile management
  - name: Products
    description: Browse and manage products
  - name: Saved Products
    description: Bookmarked / saved products (protected)
  - name: Bought Products
    description: Purchase history (protected)
  - name: Reviews
    description: Product reviews
  - name: FAQs
    description: Frequently asked questions
  - name: Polls
    description: Community polls and voting
  - name: Personal Orders
    description: Custom material orders
  - name: Analytics
    description: Product analytics (admin only)
  - name: System
    description: Health-check and diagnostic endpoints

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:          { type: integer, example: 1 }
        email:       { type: string, format: email, example: user@example.com }
        name:        { type: string, example: "John Doe" }
        avatar_url:  { type: string, nullable: true, example: "https://example.com/avatar.jpg" }
        auth_provider:
          type: string
          enum: [email, google, facebook]
        created_at:  { type: string, format: date-time }
        is_admin:    { type: boolean, default: false }

    AuthResponse:
      type: object
      properties:
        token: { type: string, example: "eyJhbGciOiJIUzI1NiJ9..." }
        user:  { $ref: '#/components/schemas/User' }

    Product:
      type: object
      properties:
        id:          { type: integer, example: 42 }
        title:       { type: string, example: "Новорічний сценарій" }
        description: { type: string }
        price:       { type: number, format: float, example: 150.00 }
        rating:      { type: number, format: float, example: 4.5 }
        type:        { type: string, example: "Сценарій" }
        image:       { type: string, example: "https://localhost:5001/uploads/products/42/cover.jpg" }
        additionalImages:
          type: array
          items: { type: string }
        ageCategories:
          type: array
          items: { type: string }
        events:
          type: array
          items: { type: string }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }

    Review:
      type: object
      properties:
        id:         { type: integer }
        productId:  { type: integer }
        userId:     { type: integer }
        userName:   { type: string }
        userAvatar: { type: string, nullable: true }
        rating:     { type: integer, minimum: 1, maximum: 5 }
        comment:    { type: string }
        createdAt:  { type: string, format: date-time }

    FAQ:
      type: object
      properties:
        id:       { type: integer }
        question: { type: string }
        answer:   { type: string }

    PersonalOrder:
      type: object
      properties:
        order_id:                   { type: integer }
        user_id:                    { type: integer }
        order_title:                { type: string }
        order_description:          { type: string }
        order_status:               { type: string }
        order_price:                { type: number }
        order_material_type:        { type: string }
        order_material_age_category: { type: string }
        order_deadline:             { type: string, format: date-time, nullable: true }
        order_created_at:           { type: string, format: date-time }

    Poll:
      type: object
      properties:
        poll_id:        { type: integer }
        poll_question:  { type: string }
        is_active:      { type: boolean }
        total_votes:    { type: integer }
        user_has_voted: { type: boolean }
        options:
          type: array
          items:
            type: object
            properties:
              vote_id:    { type: integer }
              vote_text:  { type: string }
              vote_count: { type: integer }

    Error:
      type: object
      properties:
        error:   { type: string }
        message: { type: string }
        code:    { type: string }

paths:
  # ── System ──────────────────────────────────────────────────────────────
  /health:
    get:
      tags: [System]
      summary: Health check
      description: Returns server status and current timestamp.
      responses:
        '200':
          description: Server is running
          content:
            application/json:
              example:
                status: OK
                message: "Muza Life Backend Server is running!"
                timestamp: "2026-01-01T00:00:00.000Z"
                version: "1.0.0"

  /info:
    get:
      tags: [System]
      summary: API information
      description: Returns a structured list of all available endpoints.
      responses:
        '200':
          description: API info object

  /test-db:
    get:
      tags: [System]
      summary: Database connectivity test
      responses:
        '200':
          description: Database connected
          content:
            application/json:
              example:
                status: "Database connected successfully"
                currentTime: "2026-01-01T00:00:00.000Z"
        '500':
          description: Database connection failed

  # ── Auth ─────────────────────────────────────────────────────────────────
  /auth/register/initiate:
    post:
      tags: [Auth]
      summary: Step 1 — initiate email registration
      description: |
        Validates input, checks for duplicates, generates a 6-digit OTP and
        sends it to the provided email address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:    { type: string, format: email }
                password: { type: string, minLength: 6 }
                name:     { type: string }
            example:
              email: user@example.com
              password: securepassword
              name: "John Doe"
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              example:
                success: true
                message: "Код підтвердження відправлено на вашу електронну пошту"
                email: user@example.com
                name: "John Doe"
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Email delivery error

  /auth/register/verify:
    post:
      tags: [Auth]
      summary: Step 2 — verify OTP and create user
      description: Validates OTP, hashes password, creates user record, returns JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name, verificationCode]
              properties:
                email:            { type: string, format: email }
                password:         { type: string }
                name:             { type: string }
                verificationCode: { type: string, pattern: '^\d{6}$' }
      responses:
        '201':
          description: User created; JWT returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: Invalid / expired OTP or duplicate user

  /auth/register/resend-code:
    post:
      tags: [Auth]
      summary: Resend OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: New OTP sent
        '400':
          description: User already exists

  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: |
        `loginType` controls role enforcement:
        - `'regular'` — rejects admin accounts
        - `'admin'`   — rejects non-admin accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:     { type: string, format: email }
                password:  { type: string }
                loginType:
                  type: string
                  enum: [regular, admin]
                  default: regular
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: Invalid credentials
        '403':
          description: Role / login-page mismatch

  /auth/google:
    post:
      tags: [Auth]
      summary: Login / register with Google OAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accessToken]
              properties:
                accessToken: { type: string }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid Google token

  /auth/facebook:
    post:
      tags: [Auth]
      summary: Login / register with Facebook OAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accessToken]
              properties:
                accessToken: { type: string }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid Facebook token

  # ── Products ──────────────────────────────────────────────────────────
  /products:
    get:
      tags: [Products]
      summary: Get all products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }

  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Single product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404':
          description: Product not found

  # ── Reviews ───────────────────────────────────────────────────────────
  /reviews:
    get:
      tags: [Reviews]
      summary: Get all reviews
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Review' }

  /reviews/product/{productId}:
    get:
      tags: [Reviews]
      summary: Get reviews for a product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Reviews for product

  /reviews:
    post:
      tags: [Reviews]
      summary: Create a review (protected)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, rating, comment]
              properties:
                productId: { type: integer }
                rating:    { type: integer, minimum: 1, maximum: 5 }
                comment:   { type: string }
      responses:
        '201':
          description: Review created
        '401':
          description: Unauthorized

  # ── FAQs ──────────────────────────────────────────────────────────────
  /faqs:
    get:
      tags: [FAQs]
      summary: Get all FAQs
      responses:
        '200':
          description: List of FAQ items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FAQ' }

  # ── Polls ─────────────────────────────────────────────────────────────
  /polls:
    get:
      tags: [Polls]
      summary: Get all polls
      responses:
        '200':
          description: List of polls

  /polls/{pollId}/vote:
    post:
      tags: [Polls]
      summary: Cast a vote (protected)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pollId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vote_id]
              properties:
                vote_id: { type: integer }
      responses:
        '200':
          description: Vote recorded
        '401':
          description: Unauthorized

  # ── Saved Products ────────────────────────────────────────────────────
  /saved-products:
    get:
      tags: [Saved Products]
      summary: Get saved products (protected)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saved products list
        '401':
          description: Unauthorized

  # ── Personal Orders ───────────────────────────────────────────────────
  /personal-orders:
    get:
      tags: [Personal Orders]
      summary: Get personal orders (protected)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders list
    post:
      tags: [Personal Orders]
      summary: Create personal order (protected)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderTitle, orderDescription, orderMaterialType, orderMaterialAgeCategory]
              properties:
                orderTitle:               { type: string }
                orderDescription:         { type: string }
                orderMaterialType:        { type: string }
                orderMaterialAgeCategory: { type: string }
                orderDeadline:            { type: string, format: date-time }
      responses:
        '201':
          description: Order created
        '401':
          description: Unauthorized

  # ── Analytics ─────────────────────────────────────────────────────────
  /analytics/stats/{productId}:
    get:
      tags: [Analytics]
      summary: Get product analytics (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Analytics data
        '401':
          description: Unauthorized
        '403':
          description: Admin only